moved WhereWithConditions modules in the helpers folder
