+++ spec/models/concerns/sponsor_attr_filter_spec.rb
@@ -192,41 +192,13 @@ RSpec.describe SponsorAttrFilter do
     end
   end
 
-  describe 'SponsorAttrFilter::WhereWithCondition module' do
-    describe ".where_with_conditions method" do
-      class TestClass
-        include SponsorAttrFilter::WhereWithCondition
-
-        def where(q)
-        end
-      end
-
-      let(:test_instance) {TestClass.new}
-      let(:active_relation_instance_double) {instance_double Sponsor::ActiveRecord_Relation}
-
-      it "method should exist" do
-        expect(TestClass.instance_methods.include? :where_with_conditions).to be true
-      end
-
-      it "should return an ActiveRecord_Relation when conditions are met" do
-        expect(test_instance).to receive(:where).and_return(active_relation_instance_double)
-
-        expect(test_instance.where_with_conditions([], conditions: [true])).to be active_relation_instance_double
-      end
-
-      it "should return self when conditions are not met" do
-        expect(test_instance.where_with_conditions([], conditions: [false])).to be test_instance
-      end
+  describe "should extend with WhereWithCondition module:" do
+    specify "Sponsor class methods" do
+      expect(Sponsor.methods.include? :where_with_conditions).to be true
     end
 
-    describe "should extend" do
-      specify "Sponsor class methods" do
-        expect(Sponsor.methods.include? :where_with_conditions).to be true
-      end
-
-      specify "Sponsor::ActiveRecord_Relation instance methods" do
-        expect(Sponsor::ActiveRecord_Relation.instance_methods.include? :where_with_conditions).to be true
-      end
+    specify "Sponsor::ActiveRecord_Relation instance methods" do
+      expect(Sponsor::ActiveRecord_Relation.instance_methods.include? :where_with_conditions).to be true
     end
   end
 end
